Rewrite current logic to use calcite builder
