--- modin/experimental/engines/omnisci_on_ray/frame/calcite/rel/data_type.py
@@ -1,8 +1,0 @@
-class RelDataTypeField(dict):
-    def __init__(self):
-        dict.__init__(self)
-
-
-class RelDataType(dict):
-    def __init__(self):
-        dict.__init__(self)
