--- modin/experimental/engines/omnisci_on_ray/frame/calcite/rex/correl_variable.py
@@ -1,6 +1,0 @@
-from .rex_node import RexNode
-
-
-class CorrelVariable(RexNode):
-    def __init__(self, name, typ):
-        super().__init__(name=name, type=typ)
