--- modin/experimental/engines/omnisci_on_ray/frame/calcite/rex/input_ref.py
@@ -1,6 +1,0 @@
-from .rex_node import RexNode
-
-
-class InputRef(RexNode):
-    def __init__(self, inp):
-        super().__init__(input=inp)
