--- modin/experimental/engines/omnisci_on_ray/frame/calcite/rex/literal.py
@@ -1,6 +1,0 @@
-from .rex_node import RexNode
-
-
-class Literal(RexNode):
-    def __init__(self, typ, value):
-        super().__init__(type=typ, value=value)
