--- modin/experimental/engines/omnisci_on_ray/frame/calcite/rex/local_ref.py
@@ -1,6 +1,0 @@
-from .rex_node import RexNode
-
-
-class LocalRef(RexNode):
-    def __init__(self, inp, name, typ):
-        super().__init__(input=inp, name=name, type=typ)
