--- modin/experimental/engines/omnisci_on_ray/frame/calcite/rex/rex_node.py
@@ -1,3 +1,0 @@
-class RexNode(dict):
-    def __init__(self, **kwargs):
-        dict.__init__(self, **kwargs)
